Move client related models to client-models
