+++ src/app/models/client-models/map-location.ts
