+++ src/app/views/jobs/jobs.component.ts
@@ -1,7 +1,7 @@
 import {Component, OnInit} from '@angular/core';
 import {Router} from '@angular/router';
 import {Job} from '../../models/api-models/job/job';
-import {MapLocation} from '../../models/map-location';
+import {MapLocation} from '../../models/client-models/map-location';
 import {Geolocation} from '../../services/geolocation.service';
 import {JobProxy} from '../../services/proxy/job-proxy.service';
 import {Location} from '@angular/common';
